uint64_t to uint32_t corrections

+++ src/base/eCP/ECPConfig.h
@@ -17,27 +17,27 @@ class ECPConfig
     private:
         void SetParameter(string param, string value);
 
-        uint64_t numDims;  // dimensionality of descriptors
-        uint64_t numClst;  // clusters in index
-        uint64_t numLvls;  // levels of tree index
+        uint32_t numDims;  // dimensionality of descriptors
+        uint32_t numClst;  // clusters in index
+        uint32_t numLvls;  // levels of tree index
 
-        uint64_t pgeSize;  // Size of each page
+        uint32_t pgeSize;  // Size of each page
 
         string pathIdx;  // path to index file
         string pathDat;  // data file path
 
-        uint64_t k;        // neighbors to retrieve
-        uint64_t b;        // clusters to search in
+        uint32_t k;        // neighbors to retrieve
+        uint32_t b;        // clusters to search in
 
     public:
         ECPConfig(string filename);
         ~ECPConfig();
 
-        uint64_t getNumDims()  { return numDims; };
-        uint64_t getNumClst()  { return numClst; };
-        uint64_t getNumLvls()  { return numLvls; };
+        uint32_t getNumDims()  { return numDims; };
+        uint32_t getNumClst()  { return numClst; };
+        uint32_t getNumLvls()  { return numLvls; };
 
-        uint64_t getPgeSize()  { return pgeSize; };
+        uint32_t getPgeSize()  { return pgeSize; };
 
         string getPathIdx()  { return pathIdx; };
         string getPathDat()  { return pathDat; };
